move object related msg handling to client object manager
