+++ Assets/Mirror/Runtime/ClientObjectManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 223196a76dff9054f9072d58143f698a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
